Move old database code out of way and clobber with new
